// functions/swagger/BelajarNextJsBackEnd.ts
export interface Cart {
    id?: string | undefined;
    userId?: string | undefined;
    restaurantId?: string | undefined;
    user?: User;
    restaurant?: Restaurant;
    createdAt?: Date;
}

export interface CartDetail {
    id?: string | undefined;
    cartId?: string | undefined;
    foodItemId?: string | undefined;
    qty?: number;
    cart?: Cart;
    foodItem?: FoodItem;
    createdAt?: Date;
}

export interface FoodItem {
    id?: string | undefined;
    name?: string | undefined;
    price?: number;
    restaurantId?: string | undefined;
    restaurant?: Restaurant;
    cartDetails?: CartDetail[] | undefined;
    createdAt?: Date;
}

export interface FoodItemDetailModel {
    id?: string | undefined;
    name?: string | undefined;
    price?: number;
    restaurantId?: string | undefined;
    restaurantName?: string | undefined;
}

export interface FoodItemModel {
    name?: string | undefined;
    price?: number;
    restaurantId?: string | undefined;
}

export interface Restaurant {
    id?: string | undefined;
    name?: string | undefined;
    foodItems?: FoodItem[] | undefined;
    carts?: Cart[] | undefined;
    createdAt?: Date;
}

export interface RestaurantModel {
    name?: string | undefined;
}

export interface User {
    id?: string | undefined;
    name?: string | undefined;
    email?: string | undefined;
    password?: string | undefined;
    carts?: Cart[] | undefined;
    createdAt?: Date;
}